requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
Django==5.1.7
django-allauth==65.4.1
django-crispy-forms==2.3
django-filter==25.1
django-storages==1.14.5
djangorestframework==3.16.0
Faker==37.0.0
idna==3.10
Markdown==3.8
mysqlclient==2.2.7
pillow==11.1.0
pycparser==2.22
PyJWT==2.10.1
python-dotenv==1.0.1
razorpay==1.4.2
redis==5.2.1
requests==2.32.3
setuptools==76.0.0
sqlparse==0.5.3
stripe==11.6.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
